Display the number of messages downloaded starting from one instead of zero.
Improve the behavior of the mail view in a separate window by reusing
the same size as the last resize (we could have a better policy, but
this should be enough for now).  Also, create the toolbar the right
way so that we respect the global GNOME settings without doing it
manually.

svn path=/trunk/; revision=5075