Bug 274117 – Difficult to post a new message to newsgroups
Completely rewrite how composer headers are managed.  There's now two
sets of headers (with some overlap): one set is for sending an email
message, the other is for posting to a newsgroup.  The correct set of
headers is chosen on-the-fly based on the currently selected account
type.  The user's "view" preferences for these two sets are stored
separately, so that switching between an email account and a news
account will no longer stomp on your preferences.

This also eliminates the need for the "Post New Message to Folder" and
"Post a Reply" actions, so they've been removed.  Usenet users can now
just highlight a newsgroup in the folder tree and click New or Reply,
and the composer window will show the correct set of headers with the
Post-To field filled in automatically.

Comment #6 in the bug discusses other standard news reader behaviors we
still don't get right, particularly the missing "Followup To" header.
These will be dealt with separately.