Fix libevolution-mail-shared.la
The shared library was being built as a libtool module instead of a
normal libtool library, and was installed to the wrong directory.

Fixing this created further automake-related problems which I had to
hack around (see BIG NASTY AUTOMAKE HACK in the patch).  This is NOT
a permanent solution.

A better approach is to split the source code into separate
directories, like we do in all the other components.