BUG #480095 (BNC) -  Alpha Search In Proxy Dialog Crashes Evolution.
* proxy-login.c (proxy_login_tree_view_changed_cb): Use the unique
stamp to catch only the valid iterators.

Bug description:
If you right mouse click on your Groupwise email account, you can select "Proxy
Login" and a dialog appears with all of your previously used accounts in a
list.  If you click focus into the list and then type in a keyboard letter it
crashes.  This should position you at the first match of that letter, or at
least not crash :)