Prototype EIOActivity, which integrates with GIO.
EIOActivity implements the GAsyncResult interface, and the idea is to
use this instead of GSimpleAsyncResult.  In addition to the features
offered by EActivity, it also contains GAsyncReadyCallback information
and a GCancellable.

  - Calling e_activity_cancel() triggers the GCancellable.

  - Calling e_activity_complete() triggers the GAsyncReadyCallback.

Functions that follow GIO's asynchronous pattern should return an
EIOActivity (cast as an EActivity) instead of 'void', so it can be
handed to an EShellBackend or whatever else dispatches activities.

This is not yet feature-complete.  It's missing API for storing result
values and GErrors.  I don't have a complete picture of the final API in
my head yet, so I'll copy things over from GSimpleAsyncResult as needed.