Refactor error reporting to separate error from dialogs
Previously, Most things reported errors directly.  This is evidenced by the fact
that e_error_new() returns a GtkDialog*.  This patch attempts to de-couple
error-reporting from the UI.  It introduces a simple stuct (EError) that
describes the error which is returned much like a GError by passing it as an
output parameter to a function.

e_error_new() now returns a newly-allocated EError*, but the function signature
has changed to no longer accept a parent GtkWidget, so the API change should be
detected at compile time.  I kept the convenience dialog functions, but renamed
them slightly:
 - e_error_new() -> e_error_new_dialog()
 - e_error_run() -> e_error_run_dialog()

Build is currently broken because nothing has been ported to use this new API
yet.

https://bugzilla.gnome.org/show_bug.cgi?id=602963