Make the mail folder cache a proper GObject
mail-folder-cache previously was a bit of a pseudo object (sort of a singleton)
that operated on some file static data.  This commit re-factors things so that
it is a proper class named MailFolderCache.  At the moment, this doesn't gain us
much, but in the future, it will allow us to add signals, etc so that we can
de-couple a lot of the interdependencies in here.  This is essentially a
pre-requisite to splitting up a lot of the mail backend stuff.

https://bugzilla.gnome.org/show_bug.cgi?id=604627