Add get_data_dir(), get_config_dir() vfuncs to EShellBackend
This allows modules to specify their own data dir in a flexible way without
having them hard-coded to the backend class name.  For example, the data dir for
the mail backend should be specified by the mail session (eventually as an eds
daemon) and the vfunc will allow the shell to query that in a generic way.