Save state key file asynchronously.
Avoids blocking the UI in case the save triggers an fsync() which, on
ext3, synchronizes the entire filesystem.  The logic is a bit tricky
because we want to limit the frequency of saves and only run one save
operation at a time, but we also want to do an immediate last-ditch
save before finalizing the shell view.