Add a private struct to EMailReader.
Add a private struct to the EMailReader interface so it's easier to
track state data like timeout or operation IDs, flags, etc.

This is similar to private sections of a class instance, except the
struct is created on-demand and stored as an opaque, named pointer on
the GObject with a "destroy" callback that runs during finalization.

All this is hidden behind a simple E_MAIL_READER_GET_PRIVATE macro.