[win32] Implement network status detection
The implementation is done using System Event Notification Service
by implementing ConnectionMade ConnectionMadeNoQOCInfo and
ConnectionLost methods of ISensNetwork interface.

Make the subscription to the network status notification receive
the event only if the ownerof the subscription is logged on to
the same computer as the publisher. This makes this module work
on Windows Vista and Windows 7 with normal user account.

Don't try to build Windows SENS when not building for Windows.
Extract the relevant COM structs and typedefs from mingw-w64 headers
to allow to build the module with mingw.org toolchain and fix
build breakages with Microsoft compilers.