[win32] Little modifications to Windows SENS module
Make the MySensNetwork class inherit from ISensNetwork interface
instead of IDispatch (this assures, inter alia, that we use the
right function signatures for the callbacks).

Make the subscription to the network status notification receive
the event only if the ownerof the subscription is logged on to
the same computer as the publisher. This makes this module work
on Windows Vista and Windows 7 with normal user account.