Do not hide junk mail when vJunk folder not supported by store
This is part of bug #257641, to not hide junk emails in folders
where their store doesn't support vJunk folder. In that case
the junk mail should be either placed in a real junk folder or
not having the flag set. But because there can be mails with
the junk flag set, and they would be otherwise hidden for a user,
then this allows to show them.