Utility function to hide widgets when in Express mode
Express mode requires a reduced preferences dialog.  Many options in the
current preferences dialog are pure, unadulterated crack.  So we need
an easy way to hide them, to simplify the dialog and reduce its size.

Here we add a function that takes a GConf key, reads a list of strings
from that key, and hides the widgets whose names are those strings.
This gives us an easy way to experiment with what widgets should
be hidden in the preferences dialog, without needing to recompile
all the time.

Signed-off-by: Federico Mena Quintero <federico@novell.com>