Hide EConfig sections that have no visible widgets
The EConfig code that creates widgets based on .eplug descriptions will
already hide sections that end up containing no child widgets.

Here we also make that code hide sections that end up containing
only invisible child widgets.  We will use this from the actual
plugins, so that if they decide not to show any widgets in Express
mode, then the corresponding configuration sections will not
show up in the preferences dialog.

EConfig types 'section' and 'section-table' have an internal factory function,
which doesn't return the actual GtkFrame that they create.  Instead, they return
a GtkContainer which is the actual vbox or table used to insert child widgets.

Here we modify the internal factory function to also return the actual GtkFrame
that it creates, so that the calling code can hide *that* frame properly.

Signed-off-by: Federico Mena Quintero <federico@novell.com>