Hack the row spacing of a single-row table
The table for the Time preferences in the calendar's General preferences page,
when Evo runs in Express mode, has only a single row of visible widgets;
all the other rows are hidden in Express mode.  However, GTK+ 2.20 and earlier
will include row spacings even for empty rows in GtkTable (or those that only
have invisible widgets).  Since that table has only a single visible row
in Express mode, we hack its row spacing to be 0 so that we don't get any
extra spacing in that case.

Signed-off-by: Federico Mena Quintero <federico@novell.com>