Bug 594153 (2/3) - Allow setting alarms on any meeting
The "view-time-zone" and "view-categories" actions should not have
been grouped with the "individual" action group but rather the "core"
action group, as they should always be available regardless of whether
or not the element being edited is read-only/owned, as they simply
affect visibility of various widgets.

This fixes a bug introduced in 0597b877c5bf4d21ac4048742ddf6b11e24877ba
where these two actions were accidentally grouped with other actions
that should legitimately be in the "individual" group.