Don't auto-undelete when viewing Trash folder.
When a user marks a deleted message as important or unread in a normal
folder, we automatically undelete the message.  Doing so when viewing a
virtual Trash folder makes the message immediately disappear (since it's
no longer trash), which tends to confuse and alarm users.  So limit this
behavior to normal folders.