Function to set a CamelURL's defaults based on a provider's defaults
In http://bugs.meego.com/show_bug.cgi?id=6498, part of the problem is that
when a mail account is created through the startup wizard in Express mode,
that account does not receive the default values that were defined by its
respective CamelProvider (i.e. the provider->extra_conf
CamelProviderConfEntry structures).

However, the defaults *are* used if an account is created when not in
Express mode.  The problem is that Express mode doesn't include the
"Receiving options" page in its mail account editor, while non-express
mode does.  The utility functions to populate that page's widgets are the
ones responsible for setting the provider's default values on the CamelURL
for the new account.  Since in Express mode those widgets don't even get
created, the provider's defaults are never even considered.

Here, what we do is to pull out the logic from those functions that create
widgets, so that we have set_provider_defaults_on_url(), a single function
to set default values from a CamelProvider into a CamelURL.  We will use
that function to set the defaults in both Express and non-express modes,
instead of depending on the widget code to do that.

Signed-off-by: Federico Mena Quintero <federico@novell.com>