Bug 650088 - Cannot send message from Contacts view
This starts up the EMailShellBackend whenever a new composer window is
created.  Normally this happens when switching to the Mail shell view,
but if trying to send a message from a different shell view before the
Mail shell view is ever activated, the mail accounts were not getting
loaded and sending or saving the message to a mail folder would fail.