mail-autofilter.c: Add argument guards in public functions.
This is related to bug #650587, but is not a proper fix.

This commit just trades a crash for a runtime warning, but can be
backported to the stable branch, whereas the changes required for the
proper fix are too extensive to backport.

(cherry picked from commit d5398d49ce3d2d672f8ad8c5bbb1c0f760233da5)