mail-autofilter.c: Add argument guards in public functions.
This is related to bug #650587, but is not a proper fix.  The proper fix
is in the next couple commits about EMailReader action groups.

This commit just trades a crash for a runtime warning, but can be
backported to the stable branch, whereas the changes required for the
proper fix are too extensive to backport.