Limit the size of tab labels to a reasonable size
We use half the size allocated to the whole view as a reference, and
force the label to be at most that size, taking advantage of ellipsis
otherwise. Notice that we also force the natural size as the requested
size if it is not too big as well, because labels ellipsisized using
the minimum size usually, which is undesired.

This is a fix for https://bugs.meego.com/show_bug.cgi?id=18313