EAttachmentPaned: Add "resize-toplevel" property.
Similiar to the new GtkExpander:resize-toplevel property in GTK+ 3.2,
but adapted to the fact that EAttachmentPaned's expander has no direct
child widget, and instead acts on the child widget in the lower pane.

CompEditor now uses this to fix the weird vertical resizing behavior
when its attachment bar is expanded and then collapsed again.