e-mail-store.c: Take EMailSession instead of EMailBackend.
My apologies for flip-flopping the API again.

e-mail-store.c functions used to take an EMailSession, then I changed
it to take an EMailBackend in preparation for my account-mgmt branch.

Having rethought some API decisions on the branch, however, the first
flip-flop proved to be unnecessary.  And now Srini needs the API to use
EMailSession for his mail-factory branch, so I'm flip-flopping again.