EMailBackend: Make sure all background jobs have descriptions.
All background jobs should appear in the task bar with a description.
Some CamelServices (esp. imapx) are submitting lots of background jobs
but never pushing a status message on its CamelOperation.  This commit
is to flush them out so we can find and fix them, and also so we can
clearly see what's *really* going on in the background.

All background jobs are now given an initial generic description which
should never appear in the user interface.  The callback function must
call camel_operation_push_message() to override it or else it's a bug.