Revert "Include searching Bcc header while looking for recipients"
This reverts commit 015ede7bc8212f079c3d9e2f2304b14457dbcfd9.

We can't search Bcc headers until we store Bcc addresses in the summary
database, otherwise the entire message has to be downloaded in order to
examine that one header.

Including Bcc addresses in the summary database so we can include Bcc
headers in recipient searches would be a worthy goal for 3.6.

Reverting this for now since it's a regression of bug 593020.