Rewrite em_folder_tree_model_lookup_uri() to not use uri_index.
Instead we parse the 'folder_uri' string into CamelStore and folder name
components using e_mail_folder_uri_parse().

Then we query the internal 'store_index' table with our CamelStore for
an EMFolderTreeModelStoreInfo struct.  Then we query the info struct's
'full_hash' table with our folder name for the GtkTreeRowReference.

It's a little more complicated, but it allows us to drop the internal
'uri_index' table so it's a net simplification for EMFolderTreeModel.