Make the settings capplet optional.
The evolution-settings capplet was originally designed for Anjal, it was
used in MeeGo as part of the Express Mode effort, but doesn't really fit
in GNOME 3 nowadays (nor did it really fit in GNOME 2, in my opinion).

Add a --with-capplet configure switch defaulting to 'no'.  The capplet
will eventually be removed unless I see someone actively maintaining it.